refactor: add getErrorResponse
